checkjavascript fix
Now script is checking all cells with "cell_type": "code" for the presence of plotly import. If any of them includes it, then the check is being counted as successful. Also, it checks for all outputs in a cell in case there are more than 1 present. In case "outputs" is an empty array, the iteration is skipped.