mimetype, output type analysis #5
Adding a mimetype_object and output_type_object representing the variety of mimetypes and types of output that can be in the document. By the end of checkoutput.py execution they get values based on if types inside are present or no. You can add any other types inside and the process is going to be the same.

desired_mimetypes was renamed to accepted_mimetypes and lists all mimetypes that can be in the document (has to match the mimtype_object in future, or needs to be removed at all).

Replacing if/elif with match/case statement.

Adding warnings when text/plain is detected and when notebook's size is greater than 10mb.

Adding a table with comments for each tag found in the outputs of the notebook.